FakerAndroid-As Project Path selection
